# openapi.yaml
openapi: 3.1.0
info:
  title: OpenPerpus API
  version: 0.1.0
  description: CMS Perpustakaan — 35 route handler (30 resource + health/readyz/docs/metrics/register). Deploy utama Vercel (ADR-002), Cloudflare Workers alternatif; preview lokal via `npm run dev` / `npm run cf:preview`. Enforced bidirectionally by tests/openapi-conformance.test.ts.
servers:
  - url: http://localhost:3000
    description: local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token (JWT) from client-side auth; browser sessions use cookies equivalently
  schemas:
    ErrorResponse:
      type: object
      description: Error envelope emitted by jsonError (src/lib/http-error.ts). details is a SIBLING of error, never nested inside it.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                [
                  VALIDATION,
                  NOT_FOUND,
                  CONFLICT,
                  UNAUTHORIZED,
                  FORBIDDEN,
                  THEME_FORBIDDEN,
                  HINT_USE_THEME_ENDPOINT,
                  INTERNAL,
                  SAVE_FAILED,
                  FETCH_FAILED,
                  DELETE_FAILED,
                  INVALID_JSON,
                  BAD_REQUEST,
                  SPAM_DETECTED,
                  RATE_LIMITED,
                  GONE,
                  CSRF_MISMATCH,
                  MEMBER_HAS_FINES,
                  MEMBER_OVERDUE,
                  MEMBER_LOAN_LIMIT,
                  MEMBER_CHECK_FAILED,
                ]
            message:
              type: string
        details:
          description: Optional machine-readable detail payload (sibling of error)
      required: [error]
    PaginatedResponse:
      type: object
      description: >-
        Shared list envelope used by paginated collection GETs. Single shape
        after issue #61 — `pagination` only, `meta` was removed.
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          required: [page, limit, total, totalPages]
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
      required: [data, pagination]
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
          nullable: true
        content_md:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        category:
          type: string
          nullable: true
        cover_url:
          type: string
          nullable: true
        author_id:
          type: string
          format: uuid
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
        views:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
      required: [id, title, slug, content_md, status]
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page (articles/banners/books/members/loans default 10; other collections default 20)
paths:
...
  /api/articles:
    get:
      summary: List articles (public published; draft/archived staff)
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Paginated articles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create article (staff)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content_md]
              properties:
                title:
                  type: string
                slug:
                  type: string
                  description: Optional; slugified from title when omitted or blank (auto-unique, 409 on conflict)
                excerpt:
                  type: string
                  description: 'Optional (alias: ringkasan)'
                content_md:
                  type: string
                  description: 'Markdown body (aliases: konten, content)'
                status:
                  type: string
                  enum: [draft, published, archived]
                  default: draft
                category:
                  type: string
                cover_url:
                  type: string
                  description: 'Alias: gambar_url'
      responses:
        '201':
          description: Created article
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Article'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slug already used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/categories/{id}:
    get:
      summary: Get category by id (public when active; staff otherwise)
      security:
        - BearerAuth: []
        - {}
      responses:
        '200':
          description: ok
    put:
      summary: Update category (staff)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ok
    delete:
      summary: Delete category (admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ok
...
  /api/faqs/{id}:
    get:
      summary: Get faq by id (public when active; staff otherwise)
      security:
        - BearerAuth: []
        - {}
      responses:
        '200':
          description: ok
    put:
      summary: Update faq (staff)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ok
    delete:
      summary: Delete faq (admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ok
...